Fixing tests--hopefully :)
